repactor: Cars 클래스를 사용하도록 리팩터링
